Refactor cronjob script so that drupal.log is shipped only if it exists

+++ logging/scripts/cronjob.sh
@@ -10,14 +10,19 @@
 # to a logging service.
 echo "Log shipping cronjob started ..."
 
-cd $LOGS_MOUNT_PATH || exit
+# Mount for logs must exist or exit script.
+cd /app/log || exit
 
-# Create log file for today if it doesn't already exist.
+# Script will be creating logs for today and removing yesterday's logs
 TODAY_DATE=$(date +%Y-%m-%d)
-TODAY_ACCESS_LOG=${TODAY_DATE}-access.log
-TODAY_DRUPAL_LOG=${TODAY_DATE}-drupal.log
+YESTERDAY_DATE=$(date --date="yesterday" +%Y-%m-%d)
 
-echo "> Creating log files for ${TODAY_DATE} ..."
+# Ship access.log created by nginx.
+
+# Create log file to keep track of today's access log entries.
+TODAY_ACCESS_LOG=${TODAY_DATE}-access.log
+
+echo "> Creating ${TODAY_ACCESS_LOG} ..."
 
 if [ ! -f ./${TODAY_ACCESS_LOG} ]; then
     touch ./${TODAY_ACCESS_LOG}
@@ -26,37 +31,19 @@ else
     echo "${TODAY_ACCESS_LOG} already exists"
 fi
 
-if [ ! -f ./${TODAY_DRUPAL_LOG} ]; then
-    touch ./${TODAY_DRUPAL_LOG}
-    echo "${TODAY_DRUPAL_LOG} created"
-else
-    echo "${TODAY_DRUPAL_LOG} already exists"
-fi
-
-# Compare latest access log entries with existing entries for today and add in newest entries.
+# Get latest access log entries and ship to logz.io.
 echo "> Retrieving latest log entries from /var/log/access.log and writing to ${TODAY_ACCESS_LOG}"
 cat /var/log/access.log | grep $(date +%d/%b/%Y:) > ./access-latest.log
 diff --changed-group-format='%>' --unchanged-group-format='' $TODAY_ACCESS_LOG access-latest.log > access-new.log
 cat access-new.log >> $TODAY_ACCESS_LOG
-
-# Compare latest drupal log entries with existing entries for today and add in newest entries.
-echo "> Retrieving latest log entries from /app/log/drupal.log and writing to ${TODAY_DRUPAL_LOG}"
-cat /app/log/drupal.log | grep "$(date +'%a, %d/%m/%Y -')" > ./drupal-latest.log
-diff --changed-group-format='%>' --unchanged-group-format='' $TODAY_DRUPAL_LOG drupal-latest.log > drupal-new.log
-cat drupal-new.log >> $TODAY_DRUPAL_LOG
-
-# Upload new access and drupal log entries to logz.io using cURL.
 echo "> Shipping latest log entries from /var/log/access.log to Logz.io using cURL"
 curl -T access-new.log https://listener.logz.io:8022/file_upload/${LOGZ_TOKEN}/nginx_access
-curl -T drupal-new.log https://listener.logz.io:8022/file_upload/${LOGZ_TOKEN}/drupal
 
 # Clean up temporary log files.
-rm access-latest.log access-new.log drupal-latest.log drupal-new.log
+rm access-latest.log access-new.log
 
 # Delete yesterdays log files.
-YESTERDAY_DATE=$(date --date="yesterday" +%Y-%m-%d)
 YESTERDAY_ACCESS_LOG=${YESTERDAY_DATE}-access.log
-YESTERDAY_DRUPAL_LOG=${YESTERDAY_DATE}-drupal.log
 
 echo "> Deleting yesterday's log files ..."
 
@@ -67,11 +54,43 @@ else
     echo "${YESTERDAY_ACCESS_LOG} does not exist"
 fi
 
-if [ -f ./${YESTERDAY_DRUPAL_LOG} ]; then
-    rm ${YESTERDAY_DRUPAL_LOG}
-    echo "${YESTERDAY_DRUPAL_LOG} deleted"
-else
-    echo "${YESTERDAY_DRUPAL_LOG} does not exist"
+# Ship drupal.log created by filelog module if present.
+if [ -f /app/log/drupal.log ]; then
+
+    # Create log file to keep track of today's drupal log entries.
+    TODAY_DRUPAL_LOG=${TODAY_DATE}-drupal.log
+
+    echo "> Creating ${TODAY_DRUPAL_LOG} ..."
+
+    if [ ! -f ./${TODAY_DRUPAL_LOG} ]; then
+        touch ./${TODAY_DRUPAL_LOG}
+        echo "${TODAY_DRUPAL_LOG} created"
+    else
+        echo "${TODAY_DRUPAL_LOG} already exists"
+    fi
+
+    # Get latest drupal log entries and ship logz.io.
+    echo "> Retrieving latest log entries from /app/log/drupal.log and writing to ${TODAY_DRUPAL_LOG}"
+    cat /app/log/drupal.log | grep "$(date +'%a, %d/%m/%Y -')" > ./drupal-latest.log
+    diff --changed-group-format='%>' --unchanged-group-format='' $TODAY_DRUPAL_LOG drupal-latest.log > drupal-new.log
+    cat drupal-new.log >> $TODAY_DRUPAL_LOG
+    echo "> Shipping latest drupal log entries to Logz.io using cURL"
+    curl -T drupal-new.log https://listener.logz.io:8022/file_upload/${LOGZ_TOKEN}/drupal
+
+    # Clean up temporary log files.
+    rm drupal-latest.log drupal-new.log
+
+    # Delete yesterdays log files.
+    YESTERDAY_DRUPAL_LOG=${YESTERDAY_DATE}-drupal.log
+
+    echo "> Deleting yesterday's drupal log ..."
+
+    if [ -f ./${YESTERDAY_DRUPAL_LOG} ]; then
+        rm ${YESTERDAY_DRUPAL_LOG}
+        echo "${YESTERDAY_DRUPAL_LOG} deleted"
+    else
+        echo "${YESTERDAY_DRUPAL_LOG} does not exist"
+    fi
 fi
 
 exit 0
